+++ module/Application/src/Controller/BaseControllerFactory.php
@@ -3,6 +3,7 @@
 namespace Application\Controller;
 
 use Interop\Container\ContainerInterface;
+use Zend\Authentication\AuthenticationServiceInterface;
 use Zend\Db\Adapter\AdapterInterface;
 use Zend\ServiceManager\Factory\FactoryInterface;
 
@@ -11,7 +12,8 @@ class BaseControllerFactory implements FactoryInterface
     public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
     {
         $config = $container->get('config');
+        $authService = $container->get(AuthenticationServiceInterface::class);
         $db = $container->get(AdapterInterface::class);
-        return new $requestedName($config, $db);
+        return new $requestedName($authService, $config, $db);
     }
 }