+++ module/Application/src/Module.php
@@ -3,6 +3,8 @@
 namespace Application;
 
 use Zend\Mvc\MvcEvent;
+use Zend\Session\Exception\RuntimeException as SessionValidationException;
+use Zend\Session\SessionManager;
 
 class Module
 {
@@ -19,6 +21,16 @@ class Module
             $serviceManager->get('Zend\Db\Adapter\AdapterInterface')
         );
         $errorListener->attach($eventManager);
+
+        // Create a SessionManager so that it is ready for use as the default manager for all containers.
+        $sessionManager = $serviceManager->get(SessionManager::class);
+        try {
+            // This will run the session validators so we can handle any issues centrally.
+            $sessionManager->start();
+        } catch (SessionValidationException $ex) {
+            $sessionManager->destroy();
+            $sessionManager->getValidatorChain()->clearListeners('session.validate');
+        }
     }
 
     public function getConfig()